environment: move "core.attributesfile" into repo-setting
When handling multiple repositories within the same process, relying on
global state for accessing the "core.attributesfile" configuration can
lead to incorrect values being used. It also makes it harder to isolate
repositories and hinders the libification of git.

Move the "core.attributesfile" configuration into the
`struct repo_settings` instead of relying on the global state.
Add a new function `repo_settings_get_attributesfile_path()` to retrieve
this setting in a repository-scoped manner

Reported-by: Ayush Chandekar <ayu.chandekar@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Usman Akinyemi <usmanakinyemi202@gmail.com>
Signed-off-by: Olamide Caleb Bello <belkid98@gmail.com>